requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
packaging==24.1
playwright==1.46.0
pluggy==1.5.0
pyee==11.1.0
pytest==8.3.2
pytest-base-url==2.1.0
pytest-playwright==0.5.1
python-slugify==8.0.4
requests==2.32.3
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.2
